Restore TZ hints to programmes broadcast around BST->GMT changeover.
The Radio Times feed previously included hints (at the start of
programme titles) to cater for ambiguous localtime programme start
times during the BST->GMT transition.

This patch prepends "(BST) " to programme titles starting in the hour
before 01:00 on the day of changeover, and "(GMT) " to titles starting
in the hour afterwards (after the clocks have "gone back").

Without this change, ambiguous localtimes of 01:00-01:59 are generated
in the listings and can make it impossible for downstream software to
generate the correct UTC offset for such broadcasts.