Test cases for XMLTV output BST/GMT hinting.
These test cases determine the current year's BST->GMT changeover
date/time and then check to see whether the BST/GMT title hint
should be added to the XMLTV output, and whether it is added
correctly.

Tests are included to cover a 2 hour window each side of the changeover,
which occurs at 01:00 UTC on the last Sunday in October.