Create tag uri based on alias
The tag uri used to be created based on canonical URI.
As the canonical uri is now based on numeric IDs instead
of the web-safe names, we need to use the tag alias.
This, in turn, means we need to resolve containers to
obtain their aliases when including parent refs on
children.