Stop generating hard-coded AVAttributes tag for YV TVA broadcast output
We've bumped the TVA schema version to pull in a bunch of new fields we need for channels. In the new schema
the hard-coded AVAttribtues we were generating are now invalid, so all our Broadcast transactions get
rejected by YouView. Instead of hacking this any further, YV have agreed that it's OK to just remove
the hard-coded attributes.