Amazon yv/qualities update (#263)
Closes ticket ENG-408

Under this work, there were cases where some ondemands would not be deleted. This work takes care of that by always creating DELETE requests when updating a piece of content, for the qualities that are not active. e.g. if a piece of content has SD and HD qualities (2 UPDATE ONDEMAND fragments), this now also generated 1 extra DELETE ONDEMAND fragment for UHD.