Disable startdate fudger MTS-1277 (#274)
As per the comment in the DateFudger class this was a problem
because we used to keep the hash after we've sent a delete.
We've long since fixed that as part of the amazon work. and we
now remove the delete hash when we upload payload, and remove
the payload hash when we upload delete request. As such this
workaround is no longer and can be removed. 
Discovered as part of MTS-1277.
Co-authored-by: EmiEnache <emi@metabroadcast.com>