Allow YouView Delete endpoint to accept an empty uri (#277)
This allows for specific elementIds to be removed without
knowing the uri of the content. This is helpful when the
imi has been obliterated from the content and cannot be
recreated, and used by the yv disparity fixer to remove
imis.

An attempt was made to allow the above behaviour to work
without affecting what was already happening here, even
though it appears that what is already happening is so
unintuitive it can be called wrong.